refactor MASIC QoS DSCP test
